+++ discord/raw_models.py
@@ -205,8 +205,7 @@ class RawReactionActionEvent(_RawReprMixin):
     emoji: :class:`PartialEmoji`
         The custom or unicode emoji being used.
     member: Optional[:class:`Member`]
-        The member who added the reaction. Only available if `event_type` is `REACTION_ADD`
-        and the reaction is inside a guild.
+        The member who added the reaction. Only available if the reaction occurs within a guild.
 
         .. versionadded:: 1.3